+++ src/handlers/set/src/backwardCompatibility.js
@@ -1,11 +1,11 @@
 /**
- * Function to facilitate backward compatability until v1.0.0 if fatalError setting as a boolean
+ * Function to facilitate backward compatibility until v1.0.0 if fatalError setting as a boolean
  * is provided.
  * @param {Any} inputA
  * @param {Any} inputB
  * @returns functions and settings object
  */
-const backwardCompatability = (inputA, inputB) => {
+const backwardCompatibility = (inputA, inputB) => {
   if (typeof inputA === 'boolean') {
     console.warn('Deprecation warning: The use of fatalError on set/setAll is deprecated and will be removed from version 1.0.0 onwards. Use settings object instead.');
     return {
@@ -21,4 +21,4 @@ const backwardCompatability = (inputA, inputB) => {
   };
 };
 
-module.exports = backwardCompatability;
+module.exports = backwardCompatibility;